rename `ManagedStateImpl` to `AggState`
