+++ src/stream/src/executor/managed_state/mod.rs
@@ -12,7 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub mod aggregation;
 pub mod dynamic_filter;
 pub mod join;
 pub mod top_n;
